Implemented friends reposts

+++ src/utils/utils.c
@@ -78,7 +78,7 @@ bool string_equals(string_t string1, string_t string2)
 	return strcmp(string1, string2) == 0;
 }
 
-void free_int(uint32_t *data)
+void free_int(uint16_t *data)
 {
 	free(data);
 }
