+++ src/utils/utils.h
@@ -26,7 +26,7 @@ void exit_with_error(string_t message);
 
 bool string_equals(string_t string1, string_t string2);
 
-void free_int(uint32_t *data);
+void free_int(uint16_t *data);
 
 bool compare_int(void *data1, void *data2);
